Factor out building of C++ AST into separate file.
